Refactor: integrate width and height resizer hooks
- add useResizer.ts
- remove useWidthResizer, useHeightResizer, useTriDivsWidthResizer
