+++ app/lib/hooks/index.ts
@@ -1,3 +1,1 @@
-export * from "./useHeightResizer";
-export * from "./useTriDivsWidthResizer";
-export * from "./useWidthResizer";
+export * from "./useResizer";
