+++ app/ui/picture/Detail.tsx
@@ -1,10 +1,10 @@
 import { Divider } from "@/app/ui/components";
 import PictureInfo from "@/app/ui/picture/Info";
 import PicturePreview from "@/app/ui/picture/Preview";
-import { useHeightResizer } from "@/app/lib/hooks";
+import { useResizer } from "@/app/lib/hooks";
 
 export default function PictureDetail() {
-  const { containerRef, handleMouseDown } = useHeightResizer();
+  const { containerRef, handleMouseDown } = useResizer("horizontal");
 
   return (
     <div ref={containerRef} className="flex flex-col w-full">
